backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
httpx==0.27.2
pdfplumber==0.11.4
python-docx==1.1.2
mammoth==1.8.0
python-dotenv==1.0.1
pytest==8.3.3
